Draw Thymio outline, image_to_world and world_to_image seem to be wrong

+++ src/global_map.py
@@ -298,13 +298,13 @@ def draw_path(img: np.ndarray, graph: Graph, path: list[int], source: np.ndarray
     vertices = [graph.vertices[v].astype(np.int32) for v in path]
     cv2.polylines(img, [np.array(vertices)], isClosed=False, color=(192, 64, 64), thickness=2)
 
-    cv2.line(img, source, free_source.astype(np.int32), color=(0, 0, 0), thickness=2)
+    cv2.line(img, source.astype(np.int32), free_source.astype(np.int32), color=(0, 0, 0), thickness=2)
     cv2.circle(img, free_source.astype(np.int32), color=(0, 0, 0), radius=6, thickness=-1)
-    cv2.circle(img, source, color=(0, 0, 0), radius=6, thickness=-1)
+    cv2.circle(img, source.astype(np.int32), color=(0, 0, 0), radius=6, thickness=-1)
 
-    cv2.line(img, target, free_target.astype(np.int32), color=(0, 0, 0), thickness=2)
+    cv2.line(img, target.astype(np.int32), free_target.astype(np.int32), color=(0, 0, 0), thickness=2)
     cv2.circle(img, free_target.astype(np.int32), color=(0, 0, 0), radius=6, thickness=-1)
-    cv2.circle(img, target, color=(0, 0, 0), radius=6, thickness=-1)
+    cv2.circle(img, target.astype(np.int32), color=(0, 0, 0), radius=6, thickness=-1)
 
 
 def dijkstra(adjacency_list: list[list[Edge]], source: int, target: int) -> list[int]:
